+++ source/battle-board/battle-board/lib/users/fetchCompetitionData.ts
@@ -1,0 +1,10 @@
+import CompetitionData from "@/models/interfaces/CompetitionData";
+
+export async function fetchCompetitionData(
+  competitionId: string
+): Promise<CompetitionData> {
+  const response = await fetch(
+    `${baseUrl}/api/competitions?competitionId=${competitionId}`
+  );
+  return response.json();
+}
